+++ examples/spektralwerk_core_demo.py
@@ -18,7 +18,7 @@ def main(host, port):
     spw_core = SpektralwerkCore(host=host, port=port)
 
     # request the device identity
-    identity = spw_core.get_identity()
+    identity = spw_core.get_identity().model_dump()
     print(f"device identity: {identity}")
 
     # receive wavelength array