+++ spektralwerk_scpi_client/devices/models.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel
+
+
+class Identity(BaseModel):
+    vendor: str
+    model: str
+    serial_number: str
+    firmware_version: str